requirements.txt
Flask
kafka-python
Faker